+++ src/main/java/com/secuconnect/client/api/PrepaidSalesApi.java
@@ -209,10 +209,11 @@ public class PrepaidSalesApi {
      * @param fields List of fields to include in the result, all others will be filtered out. Nested properties can be accessed with this notation: &#x60;prop1.prop2&#x60;.
      * @param q A query string to restrict the returned items to given conditions. The query string must consist of any combination of single expressions in the form &#x60;property:condition&#x60;. Property names can be nested like &#x60;property.property&#x60;.  Example: &#x60;customer.name:Meier&#x60;  A condition may contain:  * &#x60;?&#x60; as wildcard for one character;  * &#x60;*&#x60; as wildcard for any number of characters.  You can also use value ranges in the form &#x60;[min TO max]&#x60;.  Example: &#x60;customer.age:[30 TO 40]&#x60;  You can combine expressions logically by &#x60;expr AND expr&#x60; and &#x60;{expr} OR {expr}&#x60;. You can also negate an expression using &#x60;NOT {expr}&#x60;. Parenthesis &#x60;(...)&#x60; can be used to control precedence.  Example: &#x60;(NOT customer.name:meier*) AND (customer.age:[30 TO 40] OR customer.age:[50 TO 60])&#x60;
      * @param sort String with comma separated pairs of &#x60;field:order&#x60;.  Options for order:  * &#x60;asc&#x60; ascending;  * &#x60;desc&#x60; descending.
+     * @param expand Expand fields (all or a specific one)
      * @return Call to execute
      * @throws ApiException If fail to serialize the request body object
      */
-    public Call getAllCall(Integer count, Integer offset, String fields, String q, String sort) throws ApiException {
+    public Call getAllCall(Integer count, Integer offset, String fields, String q, String sort, String expand) throws ApiException {
         Object localVarPostBody = null;
 
         // create path and map variables
@@ -230,6 +231,8 @@ public class PrepaidSalesApi {
         localVarQueryParams.addAll(apiClient.parameterToPair("q", q));
         if (sort != null)
         localVarQueryParams.addAll(apiClient.parameterToPair("sort", sort));
+        if (expand != null)
+        localVarQueryParams.addAll(apiClient.parameterToPair("expand", expand));
 
         Map<String, String> localVarHeaderParams = new HashMap<String, String>();
 
@@ -252,9 +255,9 @@ public class PrepaidSalesApi {
     }
 
     @SuppressWarnings("rawtypes")
-    private Call getAllValidateBeforeCall(Integer count, Integer offset, String fields, String q, String sort) throws ApiException {
+    private Call getAllValidateBeforeCall(Integer count, Integer offset, String fields, String q, String sort, String expand) throws ApiException {
 
-        return getAllCall(count, offset, fields, q, sort);
+        return getAllCall(count, offset, fields, q, sort, expand);
     }
 
     /**
@@ -265,11 +268,12 @@ public class PrepaidSalesApi {
      * @param fields List of fields to include in the result, all others will be filtered out. Nested properties can be accessed with this notation: &#x60;prop1.prop2&#x60;.
      * @param q A query string to restrict the returned items to given conditions. The query string must consist of any combination of single expressions in the form &#x60;property:condition&#x60;. Property names can be nested like &#x60;property.property&#x60;.  Example: &#x60;customer.name:Meier&#x60;  A condition may contain:  * &#x60;?&#x60; as wildcard for one character;  * &#x60;*&#x60; as wildcard for any number of characters.  You can also use value ranges in the form &#x60;[min TO max]&#x60;.  Example: &#x60;customer.age:[30 TO 40]&#x60;  You can combine expressions logically by &#x60;expr AND expr&#x60; and &#x60;{expr} OR {expr}&#x60;. You can also negate an expression using &#x60;NOT {expr}&#x60;. Parenthesis &#x60;(...)&#x60; can be used to control precedence.  Example: &#x60;(NOT customer.name:meier*) AND (customer.age:[30 TO 40] OR customer.age:[50 TO 60])&#x60;
      * @param sort String with comma separated pairs of &#x60;field:order&#x60;.  Options for order:  * &#x60;asc&#x60; ascending;  * &#x60;desc&#x60; descending.
+     * @param expand Expand fields (all or a specific one)
      * @return PrepaidSalesList
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public PrepaidSalesList getAll(Integer count, Integer offset, String fields, String q, String sort) throws ApiException {
-        ApiResponse<PrepaidSalesList> resp = getAllWithHttpInfo(count, offset, fields, q, sort);
+    public PrepaidSalesList getAll(Integer count, Integer offset, String fields, String q, String sort, String expand) throws ApiException {
+        ApiResponse<PrepaidSalesList> resp = getAllWithHttpInfo(count, offset, fields, q, sort, expand);
         return resp.getData();
     }
 
@@ -281,11 +285,12 @@ public class PrepaidSalesApi {
      * @param fields List of fields to include in the result, all others will be filtered out. Nested properties can be accessed with this notation: &#x60;prop1.prop2&#x60;.
      * @param q A query string to restrict the returned items to given conditions. The query string must consist of any combination of single expressions in the form &#x60;property:condition&#x60;. Property names can be nested like &#x60;property.property&#x60;.  Example: &#x60;customer.name:Meier&#x60;  A condition may contain:  * &#x60;?&#x60; as wildcard for one character;  * &#x60;*&#x60; as wildcard for any number of characters.  You can also use value ranges in the form &#x60;[min TO max]&#x60;.  Example: &#x60;customer.age:[30 TO 40]&#x60;  You can combine expressions logically by &#x60;expr AND expr&#x60; and &#x60;{expr} OR {expr}&#x60;. You can also negate an expression using &#x60;NOT {expr}&#x60;. Parenthesis &#x60;(...)&#x60; can be used to control precedence.  Example: &#x60;(NOT customer.name:meier*) AND (customer.age:[30 TO 40] OR customer.age:[50 TO 60])&#x60;
      * @param sort String with comma separated pairs of &#x60;field:order&#x60;.  Options for order:  * &#x60;asc&#x60; ascending;  * &#x60;desc&#x60; descending.
+     * @param expand Expand fields (all or a specific one)
      * @return ApiResponse&lt;PrepaidSalesList&gt;
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public ApiResponse<PrepaidSalesList> getAllWithHttpInfo(Integer count, Integer offset, String fields, String q, String sort) throws ApiException {
-        Call call = getAllValidateBeforeCall(count, offset, fields, q, sort);
+    public ApiResponse<PrepaidSalesList> getAllWithHttpInfo(Integer count, Integer offset, String fields, String q, String sort, String expand) throws ApiException {
+        Call call = getAllValidateBeforeCall(count, offset, fields, q, sort, expand);
         Type localVarReturnType = new TypeToken<PrepaidSalesList>(){}.getType();
         return apiClient.execute(call, localVarReturnType);
     }
@@ -298,12 +303,13 @@ public class PrepaidSalesApi {
      * @param fields List of fields to include in the result, all others will be filtered out. Nested properties can be accessed with this notation: &#x60;prop1.prop2&#x60;.
      * @param q A query string to restrict the returned items to given conditions. The query string must consist of any combination of single expressions in the form &#x60;property:condition&#x60;. Property names can be nested like &#x60;property.property&#x60;.  Example: &#x60;customer.name:Meier&#x60;  A condition may contain:  * &#x60;?&#x60; as wildcard for one character;  * &#x60;*&#x60; as wildcard for any number of characters.  You can also use value ranges in the form &#x60;[min TO max]&#x60;.  Example: &#x60;customer.age:[30 TO 40]&#x60;  You can combine expressions logically by &#x60;expr AND expr&#x60; and &#x60;{expr} OR {expr}&#x60;. You can also negate an expression using &#x60;NOT {expr}&#x60;. Parenthesis &#x60;(...)&#x60; can be used to control precedence.  Example: &#x60;(NOT customer.name:meier*) AND (customer.age:[30 TO 40] OR customer.age:[50 TO 60])&#x60;
      * @param sort String with comma separated pairs of &#x60;field:order&#x60;.  Options for order:  * &#x60;asc&#x60; ascending;  * &#x60;desc&#x60; descending.
+     * @param expand Expand fields (all or a specific one)
      * @param callback The callback to be executed when the API call finishes
      * @return The request call
      * @throws ApiException If fail to process the API call, e.g. serializing the request body object
      */
-    public Call getAllAsync(Integer count, Integer offset, String fields, String q, String sort, final ApiCallback<PrepaidSalesList> callback) throws ApiException {
-        Call call = getAllValidateBeforeCall(count, offset, fields, q, sort);
+    public Call getAllAsync(Integer count, Integer offset, String fields, String q, String sort, String expand, final ApiCallback<PrepaidSalesList> callback) throws ApiException {
+        Call call = getAllValidateBeforeCall(count, offset, fields, q, sort, expand);
         Type localVarReturnType = new TypeToken<PrepaidSalesList>(){}.getType();
         apiClient.executeAsync(call, localVarReturnType, callback);
         return call;