+++ src/main/java/com/secuconnect/client/model/PrepaidMappingZvtResponse.java
@@ -1,0 +1,9 @@
+package com.secuconnect.client.model;
+
+/**
+* PrepaidMappingZvtResponse
+*/
+public interface PrepaidMappingZvtResponse {
+
+}
+